refactor: composite post
